+++ core/system/Factory.cpp
@@ -34,7 +34,6 @@
 #include "sph/solvers/ElasticDeformationSolver.h"
 #include "sph/solvers/EnergyConservingSolver.h"
 #include "sph/solvers/GravitySolver.h"
-#include "sph/solvers/PositionBasedSolver.h"
 #include "sph/solvers/SimpleSolver.h"
 #include "sph/solvers/StandardSets.h"
 #include "sph/solvers/SummationSolver.h"
@@ -325,8 +324,6 @@ AutoPtr<ISolver> Factory::getSolver(IScheduler& scheduler,
     case SolverEnum::DENSITY_INDEPENDENT:
         throwIfGravity();
         return makeAuto<DensityIndependentSolver>(scheduler, settings);
-    case SolverEnum::POSITION_BASED:
-        return makeAuto<PositionBasedSolver>(scheduler, settings);
     case SolverEnum::SIMPLE_SOLVER:
         throwIfGravity();
         return makeAuto<SimpleSolver>(scheduler, settings);